[Certik Audit] Fix addr len check (#300)
## Describe your changes and provide context
Previously we're adding ints with bytes, but we should cast to int and
add two ints.
## Testing performed to validate your change

+++ sei-cosmos/x/authz/keeper/keys.go
@@ -43,9 +43,9 @@ func addressesFromGrantStoreKey(key []byte) (granterAddr, granteeAddr sdk.AccAdd
 	granterAddrLen := key[1] // remove prefix key
 	kv.AssertKeyAtLeastLength(key, 3+int(granterAddrLen))
 	granterAddr = sdk.AccAddress(key[2 : 2+granterAddrLen])
-	granteeAddrLen := int(key[2+granterAddrLen])
-	kv.AssertKeyAtLeastLength(key, 4+int(granterAddrLen+byte(granteeAddrLen)))
-	granteeAddr = sdk.AccAddress(key[3+granterAddrLen : 3+granterAddrLen+byte(granteeAddrLen)])
+	granteeAddrLen := key[2+granterAddrLen]
+	kv.AssertKeyAtLeastLength(key, 4+int(granterAddrLen)+int(granteeAddrLen))
+	granteeAddr = sdk.AccAddress(key[3+granterAddrLen : 3+int(granterAddrLen)+int(granteeAddrLen)])
 
 	return granterAddr, granteeAddr
 }
